@@ -1,31 +1,40 @@
-import React, { useState, useEffect } from "react";
-import TransactionTable from "./components/TransactionTable";
-import BarChart from "./components/BarChart";
-import PieChart from "./components/PieChart";
-import { fetchStatistics, fetchBarChartData, fetchPieChartData } from "./api/api";
+// App.js
+import React, { useState } from 'react';
+import TransactionsTable from './components/TransactionsTable';
+import Statistics from './components/Statistics';
+import BarChartComponent from './components/BarChartComponent';
+import PieChartComponent from './components/PieChartComponent';
+import './App.css';
 
-const App = () => {
-  const [month, setMonth] = useState("March");
+function App() {
+  const [month, setMonth] = useState('March');
 
-  useEffect(() => {
-    fetchStatistics(month);
-    fetchBarChartData(month);
-    fetchPieChartData(month);
-  }, [month]);
+  const months = [
+    'January', 'February', 'March', 'April', 'May', 'June',
+    'July', 'August', 'September', 'October', 'November', 'December',
+  ];
 
   return (
-    <div>
+    <div className="App">
       <h1>Transactions Dashboard</h1>
-      <select value={month} onChange={(e) => setMonth(e.target.value)}>
-        {["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"].map((m) => (
-          <option key={m} value={m}>{m}</option>
-        ))}
-      </select>
-      <TransactionTable month={month} />
-      <BarChart month={month} />
-      <PieChart month={month} />
+      <div className="month-selector">
+        <label>Select Month: </label>
+        <select value={month} onChange={(e) => setMonth(e.target.value)}>
+          {months.map((m) => (
+            <option key={m} value={m}>
+              {m}
+            </option>
+          ))}
+        </select>
+      </div>
+      <Statistics month={month} />
+      <div className="charts-container">
+        <BarChartComponent month={month} />
+        <PieChartComponent month={month} />
+      </div>
+      <TransactionsTable month={month} />
     </div>
   );
-};
+}
 
 export default App;
